Update API Test 2
I wasn't sure what ""enforceSignerVisibility": true," is for but it didn't seem to affect the outcome when removing it or leaving it. Based on the definition "When true requires that a Signer have a signature or initial on the document or that the document has no signers in order to view it." I was expecting if the signer didn't have a signature on the document then a second document would not be seen, however I was seeing both documents for both signers when modified to send 2 documents.